refactor(lifecycle): unify listing state machine transition matrix
