+++ core/src/services/lifecycle/LifecycleGuard.ts
@@ -1,78 +1,18 @@
-import { LIFECYCLE_STATUS } from '@esparex/contracts';
-import { LISTING_TYPE } from '@esparex/contracts';
+import { LIFECYCLE_STATUS, LISTING_TYPE } from '@esparex/contracts';
+import {
+    ALLOWED_AD_TRANSITIONS,
+    ALLOWED_USER_TRANSITIONS,
+    ALLOWED_BUSINESS_TRANSITIONS,
+    ALLOWED_SERVICE_TRANSITIONS,
+    ALLOWED_SPARE_PART_LISTING_TRANSITIONS
+} from '@esparex/shared';
 
-/**
- * Ad Transitions
- */
-export const ALLOWED_AD_TRANSITIONS: Record<string, string[]> = {
-    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    // SOLD → DELETED is intentional: users must be able to remove sold listings from their history.
-    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    // EXPIRED → SOLD is intentional: sellers can retrospectively mark an expired listing as sold.
-    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-};
-
-/**
- * User Transitions
- */
-export const ALLOWED_USER_TRANSITIONS: Record<string, string[]> = {
-    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.SUSPENDED, LIFECYCLE_STATUS.BANNED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.SUSPENDED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.BANNED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.BANNED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.INACTIVE]: [LIFECYCLE_STATUS.LIVE],
-    [LIFECYCLE_STATUS.DELETED]: [],
-};
-
-/**
- * Business Transitions
- */
-export const ALLOWED_BUSINESS_TRANSITIONS: Record<string, string[]> = {
-    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.LIVE]: [
-        LIFECYCLE_STATUS.PENDING, // Edits force re-approval
-        LIFECYCLE_STATUS.SUSPENDED, 
-        LIFECYCLE_STATUS.EXPIRED, 
-        LIFECYCLE_STATUS.DEACTIVATED, 
-        LIFECYCLE_STATUS.SOLD, // CLOSED
-        LIFECYCLE_STATUS.DELETED
-    ],
-    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.SUSPENDED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.DELETED]: [],
-};
-
-/**
- * Service Transitions
- * NOTE: LIVE → PENDING is intentional — allows seller edits to force re-review.
- */
-export const ALLOWED_SERVICE_TRANSITIONS: Record<string, string[]> = {
-    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED],
-    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED],
-    // SOLD → DELETED is intentional: users must be able to remove sold service listings from their history.
-    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
-    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING],
-    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE],
-    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE],
-};
-
-/**
- * Spare Part Listing Transitions
- * Separate from Service transitions — spare parts can be marked SOLD.
- */
-export const ALLOWED_SPARE_PART_LISTING_TRANSITIONS: Record<string, string[]> = {
-    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED],
-    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED],
-    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING],
-    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE],
-    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE],
-    // SOLD → DELETED is intentional: users must be able to remove sold spare-part listings from their history.
-    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DELETED],
+export {
+    ALLOWED_AD_TRANSITIONS,
+    ALLOWED_USER_TRANSITIONS,
+    ALLOWED_BUSINESS_TRANSITIONS,
+    ALLOWED_SERVICE_TRANSITIONS,
+    ALLOWED_SPARE_PART_LISTING_TRANSITIONS
 };
 
 export type ValidDomain = 'ad' | 'user' | 'business' | 'service' | 'catalog_part' | 'spare_part_listing';
