+++ shared/src/stateMachines/adLifecycle.ts
@@ -1,27 +1,71 @@
-import { LISTING_STATUS, type ListingStatus } from '@esparex/contracts';
+import { LIFECYCLE_STATUS, type ListingStatus } from '@esparex/contracts';
 
 /**
  * Ad Lifecycle State Machine (SSOT)
- * Validates legal status transitions for Ads/Listings to enforce domain invariants.
+ * Validates legal status transitions across all domain entities (Ad, User, Business, Service, SparePart).
  */
-const LEGAL_AD_TRANSITIONS: Record<string, readonly string[]> = {
-    [LISTING_STATUS.DRAFT]: [LISTING_STATUS.PENDING, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.PENDING]: [LISTING_STATUS.LIVE, LISTING_STATUS.REJECTED, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.LIVE]: [LISTING_STATUS.DEACTIVATED, LISTING_STATUS.SOLD, LISTING_STATUS.EXPIRED, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.DEACTIVATED]: [LISTING_STATUS.LIVE, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.EXPIRED]: [LISTING_STATUS.LIVE, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.REJECTED]: [LISTING_STATUS.DELETED],
-    [LISTING_STATUS.SOLD]: [LISTING_STATUS.DELETED],
-    [LISTING_STATUS.DELETED]: [],
-    [LISTING_STATUS.ACTIVE]: [LISTING_STATUS.DEACTIVATED, LISTING_STATUS.SOLD, LISTING_STATUS.EXPIRED, LISTING_STATUS.DELETED],
-    [LISTING_STATUS.SUSPENDED]: [LISTING_STATUS.DELETED],
-    [LISTING_STATUS.BANNED]: [LISTING_STATUS.DELETED],
-    [LISTING_STATUS.CLOSED]: [LISTING_STATUS.DELETED],
-    [LISTING_STATUS.INACTIVE]: [LISTING_STATUS.LIVE, LISTING_STATUS.DELETED],
+export const ALLOWED_AD_TRANSITIONS: Record<string, string[]> = {
+    [LIFECYCLE_STATUS.DRAFT]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.DELETED]: [],
+    [LIFECYCLE_STATUS.ACTIVE]: [LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.SUSPENDED]: [LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.BANNED]: [LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.CLOSED]: [LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.INACTIVE]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
 };
 
-export const canTransitionAdStatus = (from: ListingStatus, to: ListingStatus): boolean => {
+export const ALLOWED_USER_TRANSITIONS: Record<string, string[]> = {
+    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.SUSPENDED, LIFECYCLE_STATUS.BANNED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.SUSPENDED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.BANNED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.BANNED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.INACTIVE]: [LIFECYCLE_STATUS.LIVE],
+    [LIFECYCLE_STATUS.DELETED]: [],
+};
+
+export const ALLOWED_BUSINESS_TRANSITIONS: Record<string, string[]> = {
+    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.LIVE]: [
+        LIFECYCLE_STATUS.PENDING,
+        LIFECYCLE_STATUS.SUSPENDED, 
+        LIFECYCLE_STATUS.EXPIRED, 
+        LIFECYCLE_STATUS.DEACTIVATED, 
+        LIFECYCLE_STATUS.SOLD,
+        LIFECYCLE_STATUS.DELETED
+    ],
+    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.SUSPENDED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.DELETED]: [],
+};
+
+export const ALLOWED_SERVICE_TRANSITIONS: Record<string, string[]> = {
+    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED],
+    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED],
+    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DEACTIVATED, LIFECYCLE_STATUS.DELETED],
+    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING],
+    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE],
+    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE],
+};
+
+export const ALLOWED_SPARE_PART_LISTING_TRANSITIONS: Record<string, string[]> = {
+    [LIFECYCLE_STATUS.PENDING]: [LIFECYCLE_STATUS.LIVE, LIFECYCLE_STATUS.REJECTED],
+    [LIFECYCLE_STATUS.LIVE]: [LIFECYCLE_STATUS.SOLD, LIFECYCLE_STATUS.EXPIRED, LIFECYCLE_STATUS.DEACTIVATED],
+    [LIFECYCLE_STATUS.REJECTED]: [LIFECYCLE_STATUS.PENDING],
+    [LIFECYCLE_STATUS.EXPIRED]: [LIFECYCLE_STATUS.PENDING, LIFECYCLE_STATUS.LIVE],
+    [LIFECYCLE_STATUS.DEACTIVATED]: [LIFECYCLE_STATUS.LIVE],
+    [LIFECYCLE_STATUS.SOLD]: [LIFECYCLE_STATUS.DELETED],
+};
+
+export const canTransitionAdStatus = (from: ListingStatus | string, to: ListingStatus | string): boolean => {
     if (from === to) return true;
-    const allowed = LEGAL_AD_TRANSITIONS[from];
+    const allowed = ALLOWED_AD_TRANSITIONS[from];
     return Array.isArray(allowed) && allowed.includes(to);
 };
